@@ -4,36 +4,39 @@
     "url": "https://github.com/cartant/eslint-plugin-rxjs-angular/issues"
   },
   "dependencies": {
-    "@typescript-eslint/experimental-utils": "^5.0.0",
+    "@typescript-eslint/utils": "^8.0.0",
     "common-tags": "^1.8.0",
-    "eslint-etc": "^5.0.0",
     "requireindex": "~1.2.0",
-    "tslib": "^2.0.0"
+    "tslib": "^2.0.0",
+    "tsutils": "^3.0.0",
+    "tsutils-etc": "^1.4.1"
   },
   "description": "ESLint rules for RxJS and Angular",
   "devDependencies": {
-    "@cartant/eslint-config": "^3.0.0",
+    "@eslint/eslintrc": "^3.0.0",
+    "@eslint/js": "^9.0.0",
     "@types/chai": "^4.2.0",
     "@types/common-tags": "^1.8.0",
-    "@types/eslint": "^8.0.0",
+    "@types/eslint": "^9.0.0",
     "@types/mocha": "^9.0.0",
     "@types/node": "^18.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
+    "@typescript-eslint/rule-tester": "^8.0.0",
     "chai": "^4.2.0",
-    "eslint": "^8.0.0",
-    "husky": "^8.0.0",
+    "copyfiles": "^2.0.0",
+    "eslint": "^9.0.0",
     "lint-staged": "^13.0.0",
     "mocha": "^9.0.0",
     "prettier": "^2.0.0",
-    "rimraf": "^3.0.0",
     "ts-node": "^10.0.0",
-    "typescript": "~4.7.4"
+    "typescript": "~4.7.4",
+    "typescript-eslint": "^8.0.0"
   },
   "files": [
     "dist",
     "docs"
   ],
-  "homepage": "https://github.com/cartant/eslint-plugin-rxjs-angular",
+  "homepage": "https://github.com/fast-facts/eslint-plugin-rxjs-angular",
   "keywords": [
     "lint",
     "rules",
@@ -45,11 +48,11 @@
   "lint-staged": {
     "*.{js,ts}": "prettier --write"
   },
-  "main": "./dist/index.js",
-  "name": "eslint-plugin-rxjs-angular",
+  "main": "./index.js",
+  "name": "eslint-plugin-rxjs-angular-updated",
   "optionalDependencies": {},
   "peerDependencies": {
-    "eslint": "^8.0.0",
+    "eslint": "^9.0.0",
     "typescript": "^4.0.0 || ^5.0.0"
   },
   "private": false,
@@ -58,20 +61,21 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/cartant/eslint-plugin-rxjs-angular.git"
+    "url": "https://github.com/fast-facts/eslint-plugin-rxjs-angular.git"
   },
   "scripts": {
+    "predist": "rm -rf dist",
     "dist": "yarn run lint && yarn run dist:build",
-    "dist:build": "yarn run dist:clean && tsc -p tsconfig-dist.json",
-    "dist:clean": "rimraf dist",
-    "lint": "eslint source/**/*.ts tests/**/*.ts",
-    "prepare": "husky install",
-    "prepublishOnly": "yarn run test && yarn run dist",
+    "dist:build": "tsc -p tsconfig.dist.json",
+    "postdist": "yarn run copy:files",
+    "copy:files": "copyfiles package.json README.md LICENSE CHANGELOG.md dist",
+    "lint": "eslint",
     "prettier": "prettier --write \"./{src,tests}/**/*.{js,json,ts,tsx}\"",
     "prettier:ci": "prettier --check \"./{src,tests}/**/*.{js,json,ts,tsx}\"",
     "test": "yarn run test:rules",
-    "test:debug": "mocha -r ts-node/register -t 5000 tests/rules/prefer-async-pipe.ts",
-    "test:rules": "mocha -r ts-node/register -t 5000 tests/rules/*.ts"
+    "test:debug": "mocha -r ts-node/register -t 5000 tests/rules/prefer-async-pipe.ts --watch-files ./**/*.ts --watch",
+    "test:rules": "mocha -r ts-node/register -t 5000 tests/rules/*.ts",
+    "test:specs": "mocha -r ts-node/register tests/*-spec.ts"
   },
-  "version": "2.0.0"
+  "version": "0.0.0-semantic-release"
 }
